\name{Gmedian-package}
\alias{Gmedian-package}
%\alias{Gmedian}
\docType{package}
\title{
\packageTitle{Gmedian}
}
\description{
The geometric median (also called spatial median or L1 median) is a robust multivariate indicator of central position. This librairy provides fast estimation procedures based on averaged stochastic gradient algorithms that can handle rapidly large samples of high dimension data. Function \code{\link{Gmedian}} computes the geometric median of a numerical data set, whereas \code{\link{GmedianCov}} computes the median covariation matrix, a useful indicator for robust PCA. Robust clustering, based on the geometric media, can also be performed with \code{\link{kGmedian}}.
%\packageDescription{Gmedian}
}
\details{

\packageDESCRIPTION{Gmedian}
\packageIndices{Gmedian}

}
\author{
\packageAuthor{Gmedian}

Maintainer: \packageMaintainer{Gmedian}
}
\references{
Cardot, H., Cenac, P. and Zitt, P-A. (2013). Efficient and fast estimation of the geometric median in Hilbert spaces with an averaged stochastic gradient algorithm. \emph{Bernoulli}, 19, 18-43.

Cardot, H. and Godichon-Baggioni, A. (2015). Fast Estimation of the Median Covariation Matrix with Application to Online Robust Principal Components Analysis.  \url{http://arxiv.org/abs/1504.02852}

Cardot, H., Cenac, P. and Monnez, J-M. (2012). A fast and recursive algorithm for clustering large datasets with k-medians.  \emph{Computational Statistics and Data Analysis}, 56, 1434-1449.
}
\keyword{Gmedian}
